New_Socio_and_Sagabz/requirements.txt
docxtpl
matplotlib
numpy
openpyxl
pandas
protobuf
python-dotenv
python_docx
tqdm